+++ src/utils/find-binary.js
@@ -65,15 +65,23 @@ export async function detectMySQLPath() {
 export async function detectPostgreSQLPath() {
     const isWin = os.platform().startsWith('win');
     const commonPaths = isWin ? [
+        'C:\\Program Files\\PostgreSQL\\17\\bin',
+        'C:\\Program Files\\PostgreSQL\\16\\bin',
         'C:\\Program Files\\PostgreSQL\\15\\bin',
         'C:\\Program Files\\PostgreSQL\\14\\bin',
         'C:\\Program Files\\PostgreSQL\\13\\bin',
+        'C:\\Program Files (x86)\\PostgreSQL\\17\\bin',
+        'C:\\Program Files (x86)\\PostgreSQL\\16\\bin',
         'C:\\Program Files (x86)\\PostgreSQL\\15\\bin'
     ] : [
         '/usr/bin',
         '/usr/local/bin',
+        '/usr/lib/postgresql/17/bin',
+        '/usr/lib/postgresql/16/bin',
         '/usr/lib/postgresql/15/bin',
-        '/usr/lib/postgresql/14/bin'
+        '/usr/lib/postgresql/14/bin',
+        '/usr/lib/postgresql/13/bin',
+        '/opt/postgresql/bin'
     ];
 
     // First try to find psql executable